Audit: validate latest results against schema
